Fixed infinite loop that cause the browser to be completely unresponsive
When finding the biggest string that will fit into the defined length, there is a possibility that the algorithm would fall in an infinite loop.

It would happen when after the first run of the algorithm, it would use a delta between current and expected.
When is goes to look for the next text delimiter and that is in the beginning of the segment being scanned, it will return same position that was passed in. Then outer cycle will add 1 to that index to avoid infinite loop, but then the $.trim(), will remove it, making the algorithm fall in a infinite loop.